requirements.txt
streamlit
langchain
langchain-groq
python-dotenv
groq
PyPDF2
langchain-community
sentence_transformers
InstructorEmbedding
faiss-cpu
